+++ apps/Conduit-UI/src/components/navigation/InnerLayouts/authenticationLayout.tsx
@@ -18,7 +18,6 @@ const AuthenticationLayout: React.FC = ({ children }) => {
 
   return (
     <StyledLayout
-      baseUrl={`${process.env.CONDUIT_URL}`}
       title={'Authentication'}
       labels={labels}
       pathNames={pathNames}