+++ apps/Conduit-UI/src/http/SecurityRequests.ts
@@ -1,13 +1,25 @@
 import axios from 'axios';
 import { CONDUIT_API } from './requestsConfig';
-import { ISecurityConfig, IPlatformTypes } from '../models/security/SecurityModels';
+import ClientPlatformEnum, {
+  ISecurityConfig,
+  IUpdateClient,
+} from '../models/security/SecurityModels';
 
 export const getAvailableClientsRequest = () => {
   return axios.get(`${CONDUIT_API}/admin/security/client`);
 };
 
-export const generateNewClientRequest = (platform: IPlatformTypes, domain?: string) => {
-  return axios.post(`${CONDUIT_API}/admin/security/client`, { platform, domain });
+export const generateNewClientRequest = (
+  platform: ClientPlatformEnum,
+  domain?: string,
+  notes?: string,
+  alias?: string
+) => {
+  return axios.post(`${CONDUIT_API}/admin/security/client`, { platform, domain, notes, alias });
+};
+
+export const updateSecurityClient = (_id: string, data: IUpdateClient) => {
+  return axios.put(`${CONDUIT_API}/admin/security/client/${_id}`, { ...data });
 };
 
 export const deleteClientRequest = (_id: string) => {