+++ apps/Conduit-UI/src/pages/database/introspection/index.tsx
@@ -1,7 +1,6 @@
 import React, { ReactElement } from 'react';
-import DatabaseLayout from '../../components/navigation/InnerLayouts/databaseLayout';
-import { Box } from '@mui/material';
-import IntrospectionLayout from '../../components/database/schemas/Introspection/IntrospectionLayout';
+import DatabaseLayout from '../../../components/navigation/InnerLayouts/databaseLayout';
+import IntrospectionLayout from '../../../components/database/schemas/Introspection/IntrospectionLayout';
 
 const Introspection = () => {
   return <IntrospectionLayout />;