+++ apps/Conduit-UI/src/pages/index.tsx
@@ -1,9 +1,18 @@
 import Head from 'next/head';
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Typography from '@mui/material/Typography';
 import Slide from '@mui/material/Slide';
 import Box from '@mui/material/Box';
-import { Container, Grid, Button, Link, Icon, useTheme, useMediaQuery } from '@mui/material';
+import {
+  Container,
+  Grid,
+  Button,
+  Link,
+  Icon,
+  useTheme,
+  useMediaQuery,
+  Divider,
+} from '@mui/material';
 import EmailIcon from '@mui/icons-material/Email';
 import LockIcon from '@mui/icons-material/Lock';
 import SchemaIcon from '@mui/icons-material/VerticalSplit';
@@ -15,6 +24,9 @@ import Swagger from '../assets/svgs/swagger.svg';
 import Image from 'next/image';
 import getConfig from 'next/config';
 import { homePageFontSizeHeader } from '../theme';
+import { useAppDispatch, useAppSelector } from '../redux/store';
+import { asyncGetIntrospectionStatus } from '../redux/slices/databaseSlice';
+import { ScreenSearchDesktopRounded } from '@mui/icons-material';
 
 const {
   publicRuntimeConfig: { CONDUIT_URL },
@@ -23,9 +35,17 @@ const {
 export const CONDUIT_API = process.env.IS_DEV ? process.env.CONDUIT_URL : CONDUIT_URL;
 
 const Home = () => {
+  const dispatch = useAppDispatch();
   const [swaggerModal, setSwaggerModal] = useState<boolean>(false);
   const theme = useTheme();
   const smallScreen = useMediaQuery(theme.breakpoints.down('sm'));
+
+  const { introspectionStatus } = useAppSelector((state) => state.databaseSlice.data);
+  useEffect(() => {
+    dispatch(asyncGetIntrospectionStatus());
+  }, [dispatch]);
+
+  console.log(introspectionStatus);
 
   return (
     <>
@@ -110,7 +130,11 @@ const Home = () => {
                 <HomePageCard
                   icon={<SecretIcon />}
                   title="set up an auth method"
-                  description="Easily login with the method of your choice!"
+                  descriptionContent={
+                    <Typography variant="subtitle2">
+                      Easily login with the method of your choice!
+                    </Typography>
+                  }
                 />
               </Link>
             </Grid>
@@ -119,8 +143,11 @@ const Home = () => {
                 <HomePageCard
                   icon={<SchemaIcon />}
                   title="create a schema"
-                  description={
-                    'Create your schema with a user friendly UI and start editing you documents right away!'
+                  descriptionContent={
+                    <Typography variant="subtitle2">
+                      Create your schema with a user friendly UI and start editing you documents
+                      right away!
+                    </Typography>
                   }
                 />
               </Link>
@@ -131,7 +158,11 @@ const Home = () => {
                 <HomePageCard
                   icon={<EmailIcon />}
                   title="set up email provider"
-                  description="Select your preferred provider and start mailing!"
+                  descriptionContent={
+                    <Typography variant="subtitle2">
+                      Select your preferred provider and start mailing!
+                    </Typography>
+                  }
                 />
               </Link>
             </Grid>
@@ -140,7 +171,42 @@ const Home = () => {
                 <HomePageCard
                   icon={<LockIcon />}
                   title="set up client secrets"
-                  description="Set up your client secret across multiple platforms!"
+                  descriptionContent={
+                    <Typography variant="subtitle2">
+                      Set up your client secret across multiple platforms!
+                    </Typography>
+                  }
+                />
+              </Link>
+            </Grid>
+            <Grid item xs={12} md={6}>
+              <Link
+                style={{ textDecoration: 'none', cursor: 'pointer' }}
+                href="/database/introspection">
+                <HomePageCard
+                  icon={<ScreenSearchDesktopRounded />}
+                  title="introspection"
+                  descriptionContent={
+                    <Box display="flex" gap={2}>
+                      <Box display="flex" flexDirection="row">
+                        <Typography>
+                          Foreign Schemas:
+                          <Typography color="error">
+                            {introspectionStatus.foreignSchemaCount}
+                          </Typography>
+                        </Typography>
+                      </Box>
+                      <Divider orientation="vertical" />
+                      <Box display="flex" flexDirection="row">
+                        <Typography>
+                          Imported Schemas:
+                          <Typography color="secondary">
+                            {introspectionStatus.importedSchemaCount}{' '}
+                          </Typography>
+                        </Typography>
+                      </Box>
+                    </Box>
+                  }
                 />
               </Link>
             </Grid>