+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.13.0-rc.0](https://github.com/ConduitPlatform/Conduit-UI/compare/v0.12.3...v0.13.0-rc.0) (2022-06-07)
+
+
+### ⚠ BREAKING CHANGES
+
+* **conduit-ui:** client id/secret validation mechanism control (#70)
+
+### Features
+
+* **conduit-ui:** client id/secret validation mechanism control ([#70](https://github.com/ConduitPlatform/Conduit-UI/issues/70)) ([669495e](https://github.com/ConduitPlatform/Conduit-UI/commit/669495ee35324566a533bd62ef03a22ef1df463e))
+* **conduit-ui:** new authentication module config options ([#69](https://github.com/ConduitPlatform/Conduit-UI/issues/69)) ([e56526e](https://github.com/ConduitPlatform/Conduit-UI/commit/e56526e72466920b7f2ff06a1879c6ceb2306479))
+* **introspection:** introspection card on homepage ofr imported/foreign schemas ([#79](https://github.com/ConduitPlatform/Conduit-UI/issues/79)) ([a02283b](https://github.com/ConduitPlatform/Conduit-UI/commit/a02283bbea29739f4959c02c4b4575f86686054e))
+* **introspection:** introspection UI ([#75](https://github.com/ConduitPlatform/Conduit-UI/issues/75)) ([19e80e6](https://github.com/ConduitPlatform/Conduit-UI/commit/19e80e68d974624caf2b35ae73f3f6b03943178c))
+* **Loading:** added a new non intrusive loading animation ([#74](https://github.com/ConduitPlatform/Conduit-UI/issues/74)) ([8caf71e](https://github.com/ConduitPlatform/Conduit-UI/commit/8caf71e6f5ad529458567dc3735089beb3bef9fa))
+
+
+### Bug Fixes
+
+* **emailValidation:** workaround for working email regexp ([#71](https://github.com/ConduitPlatform/Conduit-UI/issues/71)) ([5272fae](https://github.com/ConduitPlatform/Conduit-UI/commit/5272fae029a23048b7fabd411888aa2bb40d0988))
+* **Layout:** the config now comes from nextjs method getConfig on the client-side ([#81](https://github.com/ConduitPlatform/Conduit-UI/issues/81)) ([687c3de](https://github.com/ConduitPlatform/Conduit-UI/commit/687c3def35d80bdf1a10a7ec9f92d3aa24f812ef))
+* **Schema-editor:** responsible thunk now has the option to not show error msg when not needed ([#83](https://github.com/ConduitPlatform/Conduit-UI/issues/83)) ([cce00cc](https://github.com/ConduitPlatform/Conduit-UI/commit/cce00cc375325e84ef05fc0ec283fcab8be3eff9))
+
 ### [0.12.3](https://github.com/ConduitPlatform/Conduit-UI/compare/v0.12.2...v0.12.3) (2022-04-29)